Added preprocessor macro for windows build. Added more cmake modules

+++ src/utils.c
@@ -37,7 +37,7 @@ int load_image_jpeg(const char *path, unsigned char **dst, int *width, int *heig
         return rc;
     }
     jpg_size = file_info.st_size;
-    jpg_buffer = (unsigned char*) malloc(jpg_size + 100);
+    jpg_buffer = (unsigned char*) malloc((size_t)jpg_size + 100);
 
     int fd = open(path, O_RDONLY);
     i = 0;
